[FJ-77] charts page responsiveness
