+++ frontend/src/Components/AgencyData/AgencyData.styled.js
@@ -1,10 +1,10 @@
 import styled from 'styled-components';
 
 import MainBase from 'styles/MainBase';
-import { smallerThanTabletLandscape } from 'styles/breakpoints';
+import { phoneOnly, smallerThanTabletLandscape } from 'styles/breakpoints';
 
 export const AgencyData = styled(MainBase)`
-  overflow: hidden;
+  /* overflow: hidden; */
 `;
 
 export const ContentWrapper = styled.div`
@@ -15,5 +15,11 @@ export const ContentWrapper = styled.div`
 
   @media (${smallerThanTabletLandscape}) {
     flex-direction: column;
+    overflow-y: visible;
+  }
+
+  @media (${phoneOnly}) {
+    flex-direction: column-reverse;
+    overflow-y: hidden;
   }
 `;
