+++ frontend/src/Components/Charts/Contraband/Contraband.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import * as S from './Contraband.styled';
+import { ContrabandStyled } from './Contraband.styled';
+import * as S from 'Components/Charts/ChartSections/ChartsCommon.styled';
 
 import { useTheme } from 'styled-components';
 
@@ -85,7 +86,7 @@ function SearchRate() {
   };
 
   return (
-    <S.Contraband>
+    <ContrabandStyled>
       {renderMetaTags()}
       <S.ChartSection>
         <ChartHeader
@@ -137,7 +138,7 @@ function SearchRate() {
           </S.LegendSection>
         </S.ChartSubsection>
       </S.ChartSection>
-    </S.Contraband>
+    </ContrabandStyled>
   );
 }
