+++ frontend/src/Components/Charts/TrafficStops/TrafficStops.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import * as S from './TrafficStops.styled';
+import { TrafficStopsStyled } from './TrafficStops.styled';
+import * as S from 'Components/Charts/ChartSections/ChartsCommon.styled';
 import { useTheme } from 'styled-components';
 
 // Router
@@ -192,7 +193,7 @@ function TrafficStops() {
   const handleShareCountGraph = () => {};
 
   return (
-    <S.TrafficStops>
+    <TrafficStopsStyled>
       {/* Traffic Stops by Percentage */}
       {renderMetaTags()}
       <S.ChartSection>
@@ -270,7 +271,7 @@ function TrafficStops() {
           </S.LegendBeside>
         </S.ChartSubsection>
       </S.ChartSection>
-    </S.TrafficStops>
+    </TrafficStopsStyled>
   );
 }
