+++ frontend/src/Components/Charts/UseOfForce/UseOfForce.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import * as S from './UseOfForce.styled';
+import { UseOfForceStyled } from './UseOfForce.styled';
+import * as S from 'Components/Charts/ChartSections/ChartsCommon.styled';
 import { useTheme } from 'styled-components';
 
 // Router
@@ -117,7 +118,7 @@ function UseOfForce() {
   };
 
   return (
-    <S.UseOfForce>
+    <UseOfForceStyled>
       {renderMetaTags()}
       <S.ChartSection>
         <ChartHeader
@@ -163,7 +164,7 @@ function UseOfForce() {
           </S.PieSection>
         </S.ChartSubsection>
       </S.ChartSection>
-    </S.UseOfForce>
+    </UseOfForceStyled>
   );
 }
