+++ frontend/src/Components/Sidebar/SidebarLink.styled.js
@@ -2,16 +2,36 @@ import styled from 'styled-components';
 import { NavLink } from 'react-router-dom';
 
 import Icon from 'img/icons/Icon';
-import { smallerThanTabletLandscape } from 'styles/breakpoints';
+import { phoneOnly, smallerThanTabletLandscape } from 'styles/breakpoints';
 
 export const Wrapper = styled.div`
   position: relative;
   margin: 8px 0;
+
+  @media (${smallerThanTabletLandscape}) {
+    margin: 0;
+  }
+  @media (${phoneOnly}) {
+    border-right: 1px solid ${(p) => p.theme.colors.white};
+    flex: 1;
+    display: flex;
+    justify-content: center;
+    padding: 0.5em;
+  }
 `;
 
 export const Chevron = styled(Icon)`
   position: absolute;
   right: 105%;
+  @media (${smallerThanTabletLandscape}) {
+    right: 102%;
+    top: 50%;
+    transform: translateY(-50%);
+  }
+
+  @media (${phoneOnly}) {
+    display: none;
+  }
 `;
 
 export const SidebarLink = styled(NavLink)`
@@ -25,6 +45,11 @@ export const SidebarLink = styled(NavLink)`
   color: ${(props) => (props.active ? props.theme.colors.primaryDark : props.theme.colors.text)};
 
   @media (${smallerThanTabletLandscape}) {
-    margin-left: 1rem;
+    font-size: 16px;
+  }
+  @media (${phoneOnly}) {
+    font-size: 14px;
+    font-weight: normal;
+    line-height: 1em;
   }
 `;
