+++ frontend/src/styles/MainBase.js
@@ -4,7 +4,6 @@ export default styled.main`
   flex: 1;
   display: flex;
   flex-direction: column;
-  height: 100%;
 
   overflow-x: hidden;
   overflow-y: scroll;
